SERVER-11769 Only remove mongosniff when generating packages if it's there

+++ buildscripts/packager.py
@@ -339,7 +339,8 @@ def make_package(distro, arch, spec, srcdir):
     # Remove the mongosniff binary due to libpcap dynamic
     # linkage.  FIXME: this removal should go away
     # eventually.
-    os.unlink(sdir+("%s/usr/bin/mongosniff"%BINARYDIR))
+    if os.path.exists(sdir+("%s/usr/bin/mongosniff"%BINARYDIR)):
+      os.unlink(sdir+("%s/usr/bin/mongosniff"%BINARYDIR))
     return distro.make_pkg(arch, spec, srcdir)
 
 def make_repo(repodir):
